Macro.pm: fix expansion of nested macros, and macro escaping
In order for nested macros to work properly, and to ensure that escaping
is not removed where it shouldn’t be, we need the recursion to be
strict, i.e. never re-scan text that has already been scanned.

This was broken by the use of 3 different regexs all working on
successively evaluated text.

Fix this, and simplify the code in the process.